requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
torch
torchvision
requests
timm